+++ src/block-kit/Blocks.tsx
@@ -3,7 +3,7 @@ import { JSXSlack } from '../jsx'
 import { ArrayOutput, wrap } from '../utils'
 import { Divider, Image, Section } from './index'
 
-export interface BlockProps {
+export interface BlocksProps {
   children: JSXSlack.Children<BlockComponentProps>
 }
 
@@ -12,7 +12,7 @@ export interface BlockComponentProps {
   id?: string
 }
 
-export const Block: JSXSlack.FC<BlockProps> = props => {
+export const Blocks: JSXSlack.FC<BlocksProps> = props => {
   const normalized = wrap(props.children).map(child => {
     if (child && typeof child === 'object') {
       const isNode = (v: object): v is JSXSlack.Node =>
@@ -28,7 +28,7 @@ export const Block: JSXSlack.FC<BlockProps> = props => {
           case 'section':
             return <Section {...child.props}>{...child.children}</Section>
           default:
-            throw new Error('<Block> allows only including Block component.')
+            throw new Error('<Blocks> allows only including Block component.')
         }
       }
     }
@@ -36,4 +36,22 @@ export const Block: JSXSlack.FC<BlockProps> = props => {
   })
 
   return <ArrayOutput>{normalized}</ArrayOutput>
+}
+
+/**
+ * @deprecated `BlockProps` has been deprecated. Please use `BlocksProps`
+ *             instead.
+ */
+export type BlockProps = BlocksProps
+
+/**
+ * @deprecated A confusable `<Block>` component has been deprecated and would
+ *             remove shortly. Please use `<Blocks>` instead.
+ */
+export const Block: JSXSlack.FC<BlockProps> = props => {
+  console.warn(
+    'Deprecation warning: A confusable <Block> component has been deprecated and would remove shortly. Please use <Blocks> instead.'
+  )
+
+  return Blocks(props)
 }