+++ src/block-kit/Divider.tsx
@@ -2,7 +2,7 @@
 import { DividerBlock } from '@slack/client'
 import { JSXSlack } from '../jsx'
 import { ObjectOutput } from '../utils'
-import { BlockComponentProps } from './Block'
+import { BlockComponentProps } from './Blocks'
 
 export const Divider: JSXSlack.FC<
   BlockComponentProps & { children?: undefined }